selesai

+++ database/migrations/2025_12_10_123057_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->unsignedBigInteger('user_id');
 
             $table->date('payment_date');
-            $table->integer('amount_paid');
+            $table->unsignedBigInteger('amount_paid');
             $table->text('notes')->nullable();
 
             $table->timestamps();
